moved the main component
